Userprofilefixlogic

+++ src/main/java/com/uball/uballapp/controller/GroupController.java
@@ -17,10 +17,9 @@ public class GroupController {
         this.groupDao = groupDao;
     }
 
-
     /**
      * ounces URI accessed, displays machines page
-     * */
+     */
     @GetMapping("/creategroup")
     public String index(Model model) {
 
@@ -37,59 +36,91 @@ public class GroupController {
         }
 
 
-
     }
+
     /**
      * returns create machine form
-     * */
+     */
     @GetMapping("/groups/create")
     public String create(Model model) {
-        model.addAttribute("groupsForm", new Group());
-        return "groups/create";
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            model.addAttribute("groupsForm", new Group());
+            return "groups/create";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
+
     /**
      * accessed through create machine form and inserts into DB before redirecting to machines index page
-     * */
+     */
     @PostMapping("/groups/create")
     public String insert(@ModelAttribute Group group) {
-        groupDao.save(group);
-        return "redirect:/creategroup";
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            groupDao.save(group);
+            return "redirect:/creategroup";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
 
     /**
      * when URI accessed from machines index page, displays edit machine page by id
-     * */
+     */
     @GetMapping("/groups/{id}/edit")
     public String edit(@PathVariable long id, Model model) {
-        model.addAttribute("thisGroup", groupDao.findOne(id));
-        return "groups/edit";
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            model.addAttribute("thisGroup", groupDao.findOne(id));
+            return "groups/edit";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
 
     /**
      * accessed through edit machine form and inserts into DB before redirecting to machines index page
-     * */
+     */
     @PostMapping("/groups/{id}/edit")
     public String update(@PathVariable long id, @ModelAttribute Group group) {
-        groupDao.save(group);
-        return "redirect:/creategroup";
+
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            groupDao.save(group);
+            return "redirect:/creategroup";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
 
     @GetMapping("/groups/{id}/delete")
     public String deleteLeague(@PathVariable long id) {
-        groupDao.delete(id);
-        return "redirect:/creategroup";
+
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            groupDao.delete(id);
+            return "redirect:/creategroup";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
-
-
-
-
-
-
-
-
-
-
-
 
 
 }
